# src/config/swagger/example.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
        "version": "1.0.6",
        "title": "Swagger Petstore",
        "termsOfService": "http://swagger.io/terms/",
        "contact": { "email": "apiteam@swagger.io" },
        "license": { "name": "Apache 2.0", "url": "http://www.apache.org/licenses/LICENSE-2.0.html" }
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
...
    "schemes": ["https", "http"],
    "paths": {
...
        "/user/{username}": {
            "get": {
                "tags": ["user"],
                "summary": "Get user by user name",
                "description": "",
                "operationId": "getUserByName",
                "produces": ["application/json", "application/xml"],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The name that needs to be fetched. Use user1 for testing. ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": { "description": "successful operation", "schema": { "$ref": "#/definitions/User" } },
                    "400": { "description": "Invalid username supplied" },
                    "404": { "description": "User not found" }
                }
            },
            "put": {
                "tags": ["user"],
                "summary": "Updated user",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateUser",
                "consumes": ["application/json"],
                "produces": ["application/json", "application/xml"],
                "parameters": [
                    { "name": "username", "in": "path", "description": "name that need to be updated", "required": true, "type": "string" },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated user object",
                        "required": true,
                        "schema": { "$ref": "#/definitions/User" }
                    }
                ],
                "responses": { "400": { "description": "Invalid user supplied" }, "404": { "description": "User not found" } }
            },
            "delete": {
                "tags": ["user"],
                "summary": "Delete user",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteUser",
                "produces": ["application/json", "application/xml"],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The name that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": { "400": { "description": "Invalid username supplied" }, "404": { "description": "User not found" } }
            }
        },
...
        "/user": {
            "post": {
                "tags": ["user"],
                "summary": "Create user",
                "description": "This can only be done by the logged in user.",
                "operationId": "createUser",
                "consumes": ["application/json"],
                "produces": ["application/json", "application/xml"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": { "$ref": "#/definitions/User" }
                    }
                ],
                "responses": { "default": { "description": "successful operation" } }
            }
        }
    },
    "securityDefinitions": {
        "api_key": { "type": "apiKey", "name": "api_key", "in": "header" },
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
            "flow": "implicit",
            "scopes": { "read:pets": "read your pets", "write:pets": "modify pets in your account" }
        }
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "properties": {
                "id": { "type": "integer", "format": "int64" },
                "username": { "type": "string" },
                "firstName": { "type": "string" },
                "lastName": { "type": "string" },
                "email": { "type": "string" },
                "password": { "type": "string" },
                "phone": { "type": "string" },
                "userStatus": { "type": "integer", "format": "int32", "description": "User Status" }
            },
            "xml": { "name": "User" }
        }
    },
...
}